refactor(ingestion-slowlane): Use dash instead of underscore in filename
