+++ plugin-server/src/main/ingestion-queues/batch-processing/each-batch-ingestion.ts
@@ -5,7 +5,7 @@ import { KAFKA_EVENTS_PLUGIN_INGESTION_OVERFLOW, KAFKA_SESSION_RECORDING_EVENTS 
 import { Hub, PipelineEvent, WorkerMethods } from '../../../types'
 import { formPipelineEvent } from '../../../utils/event'
 import { status } from '../../../utils/status'
-import { ConfiguredLimiter } from '../../../utils/token_bucket'
+import { ConfiguredLimiter } from '../../../utils/token-bucket'
 import { IngestionConsumer } from '../kafka-queue'
 import { defaultConfig } from './../../../config/config'
 import { eachBatch } from './each-batch'
